made ProfileNav and added avatar to other profile pages
