+++ client/src/App.js
@@ -11,7 +11,7 @@ import TeamList from './pages/TeamList'
 import RoomListPage from './pages/RoomListPage'
 import RoomPage from './pages/RoomPage'
 import UserList from './components/UserList'
-import UserProfile from './components/UserProfile'
+import UserProfile from './pages/UserProfile'
 import UserProfilePage2 from './pages/UserProfilePage2'
 import UserProfilePage3 from './pages/UserProfilePage3'
 import FourNaughtFour from './pages/FourNaughtFour'
@@ -33,6 +33,7 @@ const App = () =>
         <Route exact path="/rooms/" render={(props) => (<RoomListPage {...props} />)} />
         <Route exact path="/rooms/:id" render={(props) => (<RoomPage {...props} />)} />
         <Route exact path="/users" render={(props) => (<UserList {...props} />)} />
+        
         <Route exact path="/users/:id" render={(props) => (<UserProfile {...props} />)} />
         <Route exact path="/users/:id/page2" render={(props) => (<UserProfilePage2 {...props} />)} />
         <Route exact path="/users/:id/page3" render={(props) => (<UserProfilePage3 {...props} />)} />
