+++ client/src/components/ProfileNav.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function ProfileNav(props) {
+
+    return (
+        <>
+            0 0 0 0
+        </>
+    )
+}
